+++ src/main/kotlin/com/vk/admstorm/diagnostic/AdmStormLogger.kt
@@ -1,8 +1,9 @@
 package com.vk.admstorm.diagnostic
 
 import com.intellij.openapi.diagnostic.DefaultLogger
+import com.vk.admstorm.services.SentryService
 
-class AdmStormLogger(delegateFactory: Factory, category: String) : DefaultLogger(category) {
+class AdmStormLogger(private val sentry: SentryService, delegateFactory: Factory, category: String) : DefaultLogger(category) {
     private val delegate = delegateFactory.getLoggerInstance(category)
 
     override fun debug(message: String) {
@@ -27,9 +28,11 @@ class AdmStormLogger(delegateFactory: Factory, category: String) : DefaultLogger
 
     override fun warn(message: String, t: Throwable?) {
         delegate.warn(message, t)
+        sentry.sendWarn(t)
     }
 
     override fun error(message: String, t: Throwable?, vararg details: String?) {
         delegate.error(message, t, *details)
+        sentry.sendError(t)
     }
 }